refactor(schemas): split TexraSettings into host-neutral core + per-host extensions
TexraSettingsSchema was a single monolith covering universal settings
(model, files, latex, toolUse, ...) and a couple of VS Code-only entries
(auth.enableVSCodeGitHub and the apiKeys.set/.remove UI command
triggers). The CLI had its own runtime fields (agent, model,
outputFormat, approvalPolicy, chat, run) declared informally as a string
array in utils/config/settingsSchema.ts.

Split the surface into layered schemas:

- src/shared/schemas/settings/coreSettings.ts — host-neutral defaults and
  shape. CoreSettingsShape is the canonical source for any host that
  could plausibly implement these settings, even if only one does today.
- src/shared/schemas/settings/vscodeSettings.ts — declares
  auth.enableVSCodeGitHub and apiKeys.* (the two settings that
  name-reference VS Code or rely on VS Code-only UI affordances).
- src/shared/schemas/settings/cliSettings.ts — formalizes the CLI
  runtime fields (agent, model, outputFormat, approvalPolicy, chat, run)
  in real Zod schemas. Used by utils/config/settingsSchema.ts to derive
  KNOWN_TEXRA_KEYS from a typed source instead of a hand-maintained
  string array.
- src/shared/schemas/settings/desktopSettings.ts — empty extension
  reserved for future Electron-only entries.

settingsConfiguration.ts now composes TexraSettingsSchema as
strictObject({ ...CoreSettingsShape, ...VscodeSettingsExtensionShape }),
and DEFAULT_TEXRA_SETTINGS as the spread of the per-layer defaults.
The parsed output is byte-identical to the pre-split monolith — every
existing consumer (flattenTexraSettings, buildTexraPackageConfiguration,
fileTypeUtils.ts, fileListingRules.ts) sees the same shape with the same
defaults.

Adds a sanity vitest verifying that the composed schema parses to the
same defaults, exposes every TEXRA_SETTING_PATH, and structurally equals
DEFAULT_TEXRA_SETTINGS. Full suite: 898 tests pass.

On-disk storage layout unchanged. The choice of "single file, namespaced
keys, pass-through unknown keys" governs future config sync work but
doesn't require any data migration today.
