+++ src/utils/config/settingsSchema.ts
@@ -1,34 +1,32 @@
 // ---------------------------------------------------------------------------
-// CLI-specific enums
-// ---------------------------------------------------------------------------
-
-export const CLI_OUTPUT_FORMATS = ['text', 'json', 'ndjson'] as const;
-export type CliOutputFormat = (typeof CLI_OUTPUT_FORMATS)[number];
-
-export const CLI_APPROVAL_POLICIES = ['never', 'ask', 'yolo'] as const;
-export type CliApprovalPolicy = (typeof CLI_APPROVAL_POLICIES)[number];
-
-// ---------------------------------------------------------------------------
 // Known `texra.*` configuration keys
 // ---------------------------------------------------------------------------
 //
 // Authoritative list of canonical `texra.*` setting keys recognized by the
-// extension and the CLI for "unknown key" detection. The structured per-key
-// schemas and defaults are the canonical source (`TEXRA_SETTING_KEYS` from
-// `@shared/schemas/settingsConfiguration`); CLI-runtime-only keys are
-// appended here so they can also pass validation.
+// extension and the CLI for "unknown key" detection.
+//
+// Two sources contribute:
+//
+// 1. The VS Code-facing settings tree (`TEXRA_SETTING_KEYS`) — derived from
+//    `TexraSettingsSchema`, which is composed of {@link CoreSettingsSchema}
+//    plus VS Code-only extensions.
+// 2. The CLI runtime fields (`CLI_SETTING_PATHS`) — declared in
+//    `@shared/schemas/settings/cliSettings`. These are top-level CLI-only
+//    keys like `agent`, `model`, `outputFormat`, etc.
 
+import { CLI_SETTING_PATHS } from '@shared/schemas/settings/cliSettings';
 import { TEXRA_SETTING_KEYS } from '@shared/schemas/settingsConfiguration';
 
-// CLI-only top-level keys (no structured schema; consumed by CLI runtime).
-const CLI_RUNTIME_KEYS = [
-  'texra.agent',
-  'texra.model',
-  'texra.outputFormat',
-  'texra.approvalPolicy',
-  'texra.chat',
-  'texra.run',
-] as const;
+// Re-export CLI enums from the canonical location so existing call sites
+// (`from '@utils/config/settingsSchema'`) keep working.
+export {
+  CLI_APPROVAL_POLICIES,
+  CLI_OUTPUT_FORMATS,
+  type CliApprovalPolicy,
+  type CliOutputFormat,
+} from '@shared/schemas/settings/cliSettings';
+
+const CLI_RUNTIME_KEYS = CLI_SETTING_PATHS.map((path) => `texra.${path}`);
 
 /** Set of all known canonical `texra.*` config keys. */
 export const KNOWN_TEXRA_KEYS: ReadonlySet<string> = new Set<string>([
